feat: Introduce new providers and refactor existing structure
- Added Gemini provider for non-interactive mode using the Gemini CLI.
- Added MacOSUI provider for interactive mode using AppleScript dialogs.
- Removed deprecated shared CLI and configuration management classes.
- Refactored project detection and utility functions into a more modular structure.
- Implemented new tests for CLI commands, configuration loading, and provider functionality.
- Ensured compatibility with existing project configurations during integration tests.
