+++ .github/workflows/release-please.yml
@@ -24,4 +24,3 @@ jobs:
         uses: googleapis/release-please-action@v4
         with:
           release-type: ruby
-          config-file: .release-please-config.json
