--- .release-please-config.json
@@ -1,14 +1,0 @@
-{
-  "packages": {
-    ".": {
-      "release-type": "ruby",
-      "package-name": "aidp",
-      "bump-minor-pre-major": true,
-      "bump-patch-for-minor-pre-major": true
-    }
-  },
-  "bootstrap-sha": "HEAD",
-  "release-type": "ruby",
-  "include-component-in-tag": false,
-  "tag-separator": "v"
-}
