+++ bin/aidp
@@ -1,5 +1,5 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
 
-require 'aidp'
-Aidp::Shared::CLI.start(ARGV)
+require "aidp"
+Aidp::CLI.start(ARGV)
