--- lib/aidp/shared/sync.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-require "fileutils"
-
-module Aidp
-  module Shared
-    # Synchronization utilities
-    class Sync
-      def self.ensure_workspace_sync
-        # Placeholder for workspace synchronization logic
-        true
-      end
-    end
-  end
-end
