+++ lib/aidp/sync.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require "fileutils"
+
+module Aidp
+  # Synchronization utilities
+  class Sync
+    def self.ensure_workspace_sync
+      # Placeholder for workspace synchronization logic
+      true
+    end
+  end
+end
