+++ lib/aidp/version.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
 module Aidp
-  module Shared
-    VERSION = "0.1.0"
-  end
+  VERSION = "0.1.0"
 end
